refactor: add prelude to minecraft_server crate
