+++ binaries/pico_limbo/src/handlers/handshake.rs
@@ -2,8 +2,7 @@ use crate::ServerState;
 use minecraft_packets::handshaking::handshake_packet::HandshakePacket;
 use minecraft_protocol::protocol_version::ProtocolVersion;
 use minecraft_protocol::state::State;
-use minecraft_server::client::Client;
-use minecraft_server::event_handler::HandlerError;
+use minecraft_server::prelude::{Client, HandlerError};
 use thiserror::Error;
 
 pub async fn on_handshake(
