+++ binaries/pico_limbo/src/handlers/status.rs
@@ -4,8 +4,7 @@ use minecraft_packets::status::ping_request_packet::PingRequestPacket;
 use minecraft_packets::status::ping_response_packet::PingResponsePacket;
 use minecraft_packets::status::status_request_packet::StatusRequestPacket;
 use minecraft_packets::status::status_response_packet::StatusResponsePacket;
-use minecraft_server::client::Client;
-use minecraft_server::event_handler::HandlerError;
+use minecraft_server::prelude::{Client, HandlerError};
 
 pub async fn on_status_request(
     state: ServerState,
