+++ crates/minecraft_server/src/connected_clients.rs
@@ -1,0 +1,4 @@
+pub trait ConnectedClients {
+    fn increment(&self);
+    fn decrement(&self);
+}
